Add freeze/thaw functionality to events when not in use
All of the events have been move to be enabled/disabled during the hide/show events.  This has the exception of the Nodegraph widget, will is always on to ensure that during destruction of the node, the editor will always be removed.  This new functionality is controlled by the \'is_frozen\' attr on the Editor widget, and all enable/disable calls have been moved under the editor widget.

Some prep work done on the directory change signal, starting prep work for adding the node_type to this, undo stack is still a little bjorked... however in theory just need to debug that, setup a full publish disp name (qlabel), and then add the node_type to the publish dir... and it would be good to go?