Changes to the Variable will now auto publish/load
These signals have been pushed into an auto load/publish scheme for a couple of different reasons.  First off, its annoying af to have to click through that every single time when we already know what the default values probably are, and that those default values can be modified if needed... Secondly, it will make it easier to handle the undo stack and move these into non undo able operations

there was a fair amount of refactoring that had to go on behind the scenes in moving a few groups around, this is only the first work in progress, and further clean up would be wise... however... From the comments made in this repo, do you really think Im wise?