Setup of Widgets2 directory
overhaul of dir structure to allow auto import from itself.  By default now, this KATANA_RESOURCES lib will be added to the sys.path arg so that it can call itself inside of Katana.

Note that Widgets is a reserved keyword in Katana for libs, so Ive forced a registry to the lib as Widgets2... =\