Moved to MultiTools
Primary work was done on restructuring the components of this library to support the MultiTools directory.  The Python Path module is now registering in the SuperTools __init__ module as that seems to be the first time that it is called and needed... will update as needed later... but preliminary tests say yay!