Added first look at is_broken to BrowserItems | Restructuring of colors module
The is_broken has been added to to the VariableManagerBrowserItem as a first look.  The handlers are still pretty borked atm and dont register correctly, but atleast the display works.  This is to mitigate the populate issue of directories not existing... as you dont necessarily just want to auto create them to make it work...

All of the colors have been moved to Utils2.colors now so that they can be accessed more easily moving forward.  The GIFs have also been rearranged into the Utils.settings module which will now act as a global settings for the entire lib