Variable Manager Browser | Populate overhaul | Fix item handlers (kinda)
overhaul to this system... checkBesterest had a flag given to it now for loading vs bypassing to stop the recursive loading that was happening while checking everything.

This has improved the fix item handlers... hopefully?

This seems to work... need to run it through its paces to see if everything is working as expected